+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 - add media rls policies migration
 - add media storage bucket upload rls policy migration
+- add media storage bucket upload trigger migration
 
 ## [0.4.0] - 2026-03-01