+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - add instagram feed media insert rls policy migration
 - add instagram feed media delete rls policy migration
 - add customers select rls policy migration
+- add logs select rls policy migration
 
 ### Changed