+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.0] - [Unreleased]
+
+### Changed
+
+#### Features
+
+##### Account
+
+- password update form to require current password
+
 ## [0.5.0] - 2026-04-06
 
 ### Added